refactored the config for docker
