+++ app/.env-example1
@@ -1,0 +1,1 @@
+GOOGLE_API_KEY=KEY
