+++ app/.env-example2
@@ -1,0 +1,6 @@
+GOOGLE_API_KEY=KEY
+POSTGRES_USER=USER!
+POSTGRES_PASSWORD=PASSWORD!
+POSTGRES_SERVER=DB_SERVER!
+POSTGRES_DATABASE=DATABASE!
+POSTGRES_PORT=5432
