--- app/configuration/.env
@@ -1,6 +1,0 @@
-GOOGLE_API_KEY=YOUR_KEY!
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=YOUR_PASSWORD!
-POSTGRES_SERVER=db
-POSTGRES_DATABASE=essensfindung
-POSTGRES_PORT=5432
